refactor: test service (#181)
@W-9126602@
